+++ src/hir/print.rs
@@ -385,6 +385,9 @@ impl PrettyPrinter {
             Stmt::Stop(e, _) => {
                 self.line(&format!("stop {}", self.expr_str(e)));
             }
+            Stmt::ScopeCancel(name, _) => {
+                self.line(&format!("stop {}", name));
+            }
             Stmt::Join(e, _) => {
                 self.line(&format!("join {}", self.expr_str(e)));
             }