+++ src/mir/lower/effects.rs
@@ -23,6 +23,14 @@ impl Lowerer {
                     *span,
                 )
             }
+            hir::Stmt::ScopeCancel(name, span) => {
+                if let Some((_, scope)) = self.scope_named.iter().rev().find(|(n, _)| n == name) {
+                    let scope = *scope;
+                    self.emit(InstKind::Call("__scope_cancel".into(), vec![scope]), Type::Void, *span)
+                } else {
+                    self.emit(InstKind::Void, Type::Void, *span)
+                }
+            }
             hir::Stmt::Join(expr, span) => {
                 let v = self.lower_expr(expr);
                 self.emit(